+++ lib/arna.dart
@@ -56,4 +56,5 @@ export 'package:arna/src/theme/text_theme.dart';
 export 'package:arna/src/theme/theme.dart';
 export 'package:arna/src/theme/theme_data.dart';
 export 'package:flutter/material.dart' show Icons;
+export 'package:flutter/services.dart' show Brightness;
 export 'package:flutter/widgets.dart';