+++ lib/src/theme/colors.dart
@@ -2,7 +2,6 @@ import 'dart:math' as math;
 
 import 'package:arna/arna.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart' show Brightness;
 
 /// A palette of [Color] constants that describe colors
 class ArnaColors {