+++ lib/src/theme/theme_data.dart
@@ -1,6 +1,5 @@
 import 'package:arna/arna.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart' show Brightness;
 
 const _ArnaThemeDefaults _kDefaultTheme = _ArnaThemeDefaults(
   null,